+++ Trading-Api/code/src/main/java/service/payment/PaymentAdapter.java
@@ -2,5 +2,5 @@ package service.payment;
 
 public interface PaymentAdapter {
 
-    public void makePurchase(String accountNumber , int amount) throws Exception;
+    public void makePurchase(String accountNumber , double amount) throws Exception;
 }
